requirements.txt
#Python 3.12.3
#env name==movie_review
#streamlit run streamlit_app.py
absl-py==2.1.0
altair==5.5.0
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==25.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
Django==5.1.3
executing==2.0.1
Flask==3.0.3
flatbuffers==24.12.23
fonttools==4.53.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h5py==3.12.1
httplib2==0.30.0
idna==3.10
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.7.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mysql-connector-python==8.4.0
namex==0.0.8
narwhals==2.2.0
nest-asyncio==1.6.0
nltk==3.4.5
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.1
pywin32==306
pyzmq==26.0.3
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.27.1
rsa==4.9.1
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
setuptools==75.6.0
six==1.16.0
smmap==5.0.2
sqlparse==0.5.1
stack-data==0.6.3
streamlit==1.49.1
tenacity==9.1.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.5.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.2.0
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.45.1
wrapt==1.17.0
